# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hschool API",
        "version": "1.0.0"
    },
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "Completed successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "surname": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Completed successfully"
                    }
                }
            }
        },
        "/user/{_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "name": "Enter user id",
                        "in": "path",
                        "description": "user ID",
                        "required": true,
                        "schemas": {
                            "id": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Completed successfully"
                    }
                }
            }
        }
    }
}